modified:   __pycache__/config.cpython-311.pyc modified:   __pycache__/sql.cpython-311.pyc modified:   "assets/main/\320\263\320\273\320\260\320\262\320\275\320\260\321\217.html" modified:   assets/shop/shop.html modified:   bd.db modified:   bot.py modified:   config.py modified:   main.py modified:   sql.py modified:   static/Scripts/shop_load.js modified:   static/css/css.css modified:   static/css/shop.css new file:   static/img/profile.png new file:   "static/img/\320\272\320\276\320\270\320\2752.png" new file:   "static/img/\321\204\320\276\320\2753.png"

+++ bot.py
@@ -11,22 +11,32 @@ storage = MemoryStorage()
 dp = Dispatcher(bot, storage=storage)
 db = Database("bd.db")
 from aiogram.types import InlineKeyboardButton,InlineKeyboardMarkup, WebAppInfo,BotCommand
-@dp.message_handler(commands=['start'])
+@dp.message_handler(commands=['start','open'])
 async def bot_start(message: types.Message):
     if message.chat.type == 'private':
-        await bot.delete_message(message.chat.id, message.message_id) 
-        if not db.user_exists( message.from_user.id):
-            db.add_user(message.from_user.id)
-        ikb_donate = InlineKeyboardMarkup(row_width=1,
-                                inline_keyboard=[
-                                    [
-                                        InlineKeyboardButton(text='Перейти в магазин', web_app=WebAppInfo(url=config.main_link))
-                                    ]
-                                ])
-        await bot.send_message(message.from_user.id, 
-            f'Привет! Я бот для продажи мерча кампуса НЕЙМАРКА. Здесь вы можете найти уникальные товары, связанные с нашим брендом, включая одежду, аксессуары и сувениры.\n'
-            f'Запусти меня, чтобы увидеть весь ассортимент!',  
-        reply_markup=ikb_donate)
+        if db.checkRules(message.from_user.id):
+            await bot.delete_message(message.chat.id, message.message_id) 
+            if not db.user_exists( message.from_user.id):
+                db.add_user(message.from_user.id)
+            ikb_donate = InlineKeyboardMarkup(row_width=1,
+                                    inline_keyboard=[
+                                        [
+                                            InlineKeyboardButton(text='Перейти в магазин', web_app=WebAppInfo(url=config.main_link))
+                                        ]
+                                    ])
+            await bot.send_message(message.from_user.id, config.helloText,reply_markup=ikb_donate)
+        else:
+            confirm = InlineKeyboardMarkup(row_width=1,
+                                    inline_keyboard=[
+                                        [
+                                            InlineKeyboardButton(text='Соглашаюсь', callback_data="✅ConfirmRules")
+                                        ]
+                                    ])
+            await bot.send_message(message.from_user.id, 
+                config.helloText,  
+            reply_markup=confirm)
+            
+            pass
 async def set_default_commands(dp):
     await dp.bot.set_my_commands(
         [
@@ -35,6 +45,23 @@ async def set_default_commands(dp):
     )
 async def on_startup(dispatcher):
     await set_default_commands(dispatcher)
+@dp.callback_query_handler( text_contains = "✅ConfirmRules")
+async def Status_change(call: types.CallbackQuery):
+    db.set_rules(1,call.message.chat.id)	
+    await call.answer()
+    try:
+        await call.message.delete()
+    except:
+        pass
+    ikb_donate = InlineKeyboardMarkup(row_width=1,
+                                    inline_keyboard=[
+                                        [
+                                            InlineKeyboardButton(text='Перейти в магазин', web_app=WebAppInfo(url=config.main_link))
+                                        ]
+                                    ])
+    
+    await bot.send_message(call.message.chat.id, config.helloText, 
+                    reply_markup = ikb_donate)
 if __name__ == '__main__':
     loop = asyncio.new_event_loop()
     asyncio.set_event_loop(loop)
